analyzer: Move curation classes to their own package
To group classes with more upcoming provider implementations for a
better overview.
