+++ analyzer/src/main/kotlin/Analyzer.kt
@@ -19,6 +19,7 @@
 
 package com.here.ort.analyzer
 
+import com.here.ort.analyzer.curation.FilePackageCurationProvider
 import com.here.ort.analyzer.managers.Unmanaged
 import com.here.ort.downloader.VersionControlSystem
 import com.here.ort.downloader.vcs.GitRepo
