+++ analyzer/src/main/kotlin/curation/FilePackageCurationProvider.kt
@@ -17,8 +17,9 @@
  * License-Filename: LICENSE
  */
 
-package com.here.ort.analyzer
+package com.here.ort.analyzer.curation
 
+import com.here.ort.analyzer.PackageCurationProvider
 import com.here.ort.model.Identifier
 import com.here.ort.model.PackageCuration
 import com.here.ort.model.OutputFormat
