+++ analyzer/src/test/kotlin/curation/FilePackageCurationProviderTest.kt
@@ -17,7 +17,7 @@
  * License-Filename: LICENSE
  */
 
-package com.here.ort.analyzer
+package com.here.ort.analyzer.curation
 
 import com.here.ort.model.Identifier
